worker/requirements.txt
pandas==2.2.0
geopandas==0.14.3
fsspec==2024.2.0
SQLAlchemy==2.0.25
geoalchemy2==0.14.4
psycopg2-binary==2.9.9
requests==2.31.0
gtfs-realtime-bindings==1.0.0
kafka-python==2.0.2